backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langgraph==0.2.62
langchain-anthropic==0.3.3
langchain-core==0.3.28
pydantic==2.10.4
python-dotenv==1.0.1
slowapi==0.1.9
